How does the app start? the "administrator" button
When the app starts first time, it asks for the user to create the
first admin account.

After the first time of use, the app just opens the “DisabilityKiosk”
window.

I already created an account: user name “viet”, and password “123456”.
We can use this root account to create another admin account as well as
do all admin tasks.