Added more requirements
I moved the information regarding the administrator from the "usability" section to the "functionality" section because usability has to do with how easy/difficult it is for users to use/misuse the system (not how they are able to use it-- I believe that is what the functionality section is for). I also added information about adding/editting/removing specialists.